# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DAWNSTAR Family Health Keeper",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/backup/export": {
      "post": {
        "tags": [
          "Backup & Restore"
        ],
        "summary": "Export Backup",
        "description": "Export all household data as a downloadable ZIP archive.",
        "operationId": "export_backup_api_v1_backup_export_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backup/validate": {
      "post": {
        "tags": [
          "Backup & Restore"
        ],
        "summary": "Validate Backup",
        "description": "Validate a backup archive and stage it for import.",
        "operationId": "validate_backup_api_v1_backup_validate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_validate_backup_api_v1_backup_validate_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupValidationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backup/import": {
      "post": {
        "tags": [
          "Backup & Restore"
        ],
        "summary": "Import Backup",
        "description": "Import a previously validated backup archive.",
        "operationId": "import_backup_api_v1_backup_import_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupImportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backup/staging/{validation_id}": {
      "delete": {
        "tags": [
          "Backup & Restore"
        ],
        "summary": "Cleanup Staging",
        "description": "Clean up a staged backup file.",
        "operationId": "cleanup_staging_api_v1_backup_staging__validation_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "validation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Validation Id"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BackupCounts": {
        "properties": {
          "members": {
            "type": "integer",
            "title": "Members",
            "default": 0
          },
          "providers": {
            "type": "integer",
            "title": "Providers",
            "default": 0
          },
          "provider_assignments": {
            "type": "integer",
            "title": "Provider Assignments",
            "default": 0
          },
          "health_records": {
            "type": "integer",
            "title": "Health Records",
            "default": 0
          },
          "attachments": {
            "type": "integer",
            "title": "Attachments",
            "default": 0
          },
          "ai_insights": {
            "type": "integer",
            "title": "Ai Insights",
            "default": 0
          },
          "conversations": {
            "type": "integer",
            "title": "Conversations",
            "default": 0
          },
          "messages": {
            "type": "integer",
            "title": "Messages",
            "default": 0
          },
          "reminders": {
            "type": "integer",
            "title": "Reminders",
            "default": 0
          },
          "notifications": {
            "type": "integer",
            "title": "Notifications",
            "default": 0
          }
        },
        "type": "object",
        "title": "BackupCounts"
      },
      "BackupImportRequest": {
        "properties": {
          "validation_id": {
            "type": "string",
            "title": "Validation Id"
          },
          "mode": {
            "type": "string",
            "enum": [
              "merge",
              "replace"
            ],
            "title": "Mode"
          }
        },
        "type": "object",
        "required": [
          "validation_id",
          "mode"
        ],
        "title": "BackupImportRequest"
      },
      "BackupImportResponse": {
        "properties": {
          "imported": {
            "$ref": "#/components/schemas/BackupCounts"
          },
          "skipped": {
            "$ref": "#/components/schemas/BackupCounts"
          },
          "failed": {
            "type": "integer",
            "title": "Failed",
            "default": 0
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "imported",
          "skipped"
        ],
        "title": "BackupImportResponse"
      },
      "BackupManifest": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version",
            "default": "1.0"
          },
          "app_version": {
            "type": "string",
            "title": "App Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "household_name": {
            "type": "string",
            "title": "Household Name"
          },
          "household_id": {
            "type": "string",
            "format": "uuid",
            "title": "Household Id"
          },
          "counts": {
            "$ref": "#/components/schemas/BackupCounts"
          }
        },
        "type": "object",
        "required": [
          "app_version",
          "created_at",
          "household_name",
          "household_id",
          "counts"
        ],
        "title": "BackupManifest"
      },
      "BackupValidationResponse": {
        "properties": {
          "validation_id": {
            "type": "string",
            "title": "Validation Id"
          },
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "manifest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BackupManifest"
              },
              {
                "type": "null"
              }
            ]
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "default": []
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "validation_id",
          "valid"
        ],
        "title": "BackupValidationResponse"
      },
...
      "Body_validate_backup_api_v1_backup_validate_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_validate_backup_api_v1_backup_validate_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}